fix nil-pointer dereference, wait for VFs
Fix nil-pointer dereference when destroying a PCIDevice(Claim) for a
PCIDevice where PCIPassthrough is already disabled.

SRIOVNetworkDevice: Wait for VFs before updating ready status.
Wait for all PCIDevices corresponding to the VFs of the
SRIOVNetworkDevice to show up before proclaiming the SRIOVNetworkDevice
as ready.
This allows better ordering of operations when enabling VFs.

+++ internal/provider/pcidevice/resource_pcidevice.go
@@ -124,13 +124,14 @@ func resourcePCIDeviceDelete(ctx context.Context, d *schema.ResourceData, meta i
 	if err != nil {
 		return diag.FromErr(err)
 	}
-	err = c.HarvesterDeviceClient.DevicesV1beta1().PCIDeviceClaims().Delete(ctx, toDelete.(*devicesv1.PCIDeviceClaim).Name, metav1.DeleteOptions{})
-	if err != nil {
-		if !apierrors.IsNotFound(err) {
-			return diag.FromErr(err)
+	if toDelete.(*devicesv1.PCIDeviceClaim) != nil {
+		err = c.HarvesterDeviceClient.DevicesV1beta1().PCIDeviceClaims().Delete(ctx, toDelete.(*devicesv1.PCIDeviceClaim).Name, metav1.DeleteOptions{})
+		if err != nil {
+			if !apierrors.IsNotFound(err) {
+				return diag.FromErr(err)
+			}
 		}
 	}
-
 	d.SetId("")
 	return diag.FromErr(resourcePCIDeviceWaitForState(ctx, d, meta, schema.TimeoutDelete))
 }
